\name{acaule}

\alias{acaule}

\docType{data}

\title{Solanum acaule data}

\description{
Distribution data for Solanum acaule (a plant species that occurs in the high Andes of Peru and Bolivia). 
Downloaded from GBIF with the \code{\link[dismo]{gbif}} function. For use in the 'species distribution modeling' vignette.
}

\usage{
data(acaule)
}

\references{
\url{http://www.gbif.org}
}

\keyword{datasets}

