\name{biogeomancer}

\alias{biogeomancer}

\title{ georeferencing}

\description{
 georeferencing
}

\usage{
biogeomancer(country = "", adm1 = "", adm2 = "", locality = "")
}

\arguments{
  \item{country}{ }
  \item{adm1}{ }
  \item{adm2}{ }
  \item{locality}{}
}

\value{
data frame
}

\references{ 
Guralnick, R.P., J. Wieczorek, R. Beaman, R.J. Hijmans and the BioGeomancer Working Group, 2006. BioGeomancer:  Automated georeferencing to map the world's biodiversity data. PLoS Biology 4: 1908-1909.
\url{http://dx.doi.org/10.1371/journal.pbio.0040381}

Wieczorek, J., Q. Guo and R.J. Hijmans, 2004. The point-radius method for georeferencing point localities and calculating associated uncertainty. 
International Journal of Geographic Information Science 18: 745-767.

\url{http://bg.berkeley.edu/latest/}
}

\author{Robert J. Hijmans}


\keyword{ spatial }

