\name{boxplot}

\docType{methods}

\alias{boxplot,ModelEvaluation-method}


\title{
Box plot of model evaluation data 
}

\description{
Make a box plot of model evaluation data, i.e., the model predictions for known presence and absence points.
}


\details{
Arguments:

\code{x} Object of class ModelEvaluation
\code{...} Additional arguments that can be passed to \link[graphics]{boxplot}
}


\seealso{
\code{\link[dismo]{evaluate} }
}

\author{
Robert J. Hijmans
}


\keyword{spatial}

