\name{plots}

\alias{density}
\alias{density,DistModel-method}
\alias{pairs}
\alias{pairs,DistModel-method}

\title{density}

\description{
density and pair plots
}

\usage{
density(x, ...)
pairs(x, ...)
}

\arguments{
\item{x}{Object of class MaxEnt}
\item{...}{Additional arguments. See Details}
}

\details{
tbd
}

\author{Robert J. Hijmans}


\keyword{spatial}
