\name{dismo-package}

\alias{dismo-package}
\alias{dismo}

\docType{package}

\title{
Species distribution modeling
}

\description{
This package implements a few species distribution models, including an R link to maxent, and native implementations of older models Bioclim and Domain. It also provides a number of funcitons useful for other methods available in R such as GLM, GAM, Random Forest, and Boosted Regresssion Trees.
}

\details{
\tabular{ll}{
Package: \tab dismo\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 18-Jan-2010\cr
License: \tab GPL 3\cr
LazyLoad: \tab yes\cr
}
}

\author{
Robert J. Hijmans and Steven Phillips
}

\keyword{ package }
\keyword{ spatial }

