\name{gbif}
\alias{gbif}

\title{
Download species occurence records from the GBIF data portal.
}

\description{
.. description ...

}


\usage{
gbif(genus, species="", geo=TRUE, sp=FALSE, removeZeros=TRUE, download=TRUE, getAlt=TRUE, feedback=3)
}
 
\arguments{
  \item{genus}{ }
  \item{species}{}
  \item{geo}{}
  \item{sp}{}
  \item{removeZeros}{}
  \item{download}{}
  \item{getAlt}{}
  \item{feedback}{}
}
\details{

}


\value{
data frame 
}


\references{
\url{http://data.gbif.org/occurrences/}
}

\author{
Robert J. Hijmans
}


\keyword{spatial}

