\name{mahal}
\alias{mahal}
\alias{mahal,Raster,Spatial-method}
\alias{mahal,Raster,matrix-method}
\alias{mahal,Raster,data.frame-method}
\alias{mahal,matrix,missing-method}
\alias{mahal,data.frame,missing-method}


\title{
Mahalanobis model
}

\description{
.. description ...
}

\usage{
mahal(x, p, ...)
}
 
\arguments{
  \item{x}{ Raster* object or matrix}
  \item{p}{ two column matrix or SpatialPoints* object}
  \item{...}{ Additional arguments }
}

\details{
}


\value{
A RasterLayer or (if x is a matrix), a vector.
}


\author{
Robert J. Hijmans
}


\keyword{spatial}

