\name{pairs}

\docType{methods}

\alias{pairs}
\alias{pairs,DistModel-method}


\title{
plot pairs of occurence data
}

\usage{
pairs(x, ...)
}

\arguments{
\item{x}{Object of class DistModel or derived from that class (such as 'MaxEnt', 'Bioclim')}
\item{...}{Additional arguments. See Details}
}


\description{
.. description ...
}


\section{Methods}{
\describe{
	\code{pairs(x, ...)}
}
}


\seealso{
\code{\link[dismo]{density}, \link[dismo]{plot}}
}

\author{
Robert J. Hijmans
}


\keyword{spatial}

