\name{plot}

\docType{methods}

\alias{plot,ConfusionMatrix,ANY-method}
\alias{plot,Bioclim,missing-method}
\alias{plot,Bioclim,numeric-method}
\alias{points,Bioclim-method}


\title{
plot
}

\description{
.. description ...
}


\section{Methods}{
\describe{
 
\code{plot(x, ...)}
  
}
}


\author{
Robert J. Hijmans
}


\keyword{spatial}

