\name{pointValues}
\alias{pointValues}

\title{
point values
}


\description{
.. description ...
}


\usage{
pointValues(stack, p, a, uniquecells = TRUE, na.rm = TRUE)
}
 
\arguments{
  \item{stack}{}
  \item{p}{}
  \item{a}{}
  \item{uniquecells}{}
  \item{na.rm}{}
}



\value{
}


\author{
Robert J. Hijmans
}


\keyword{spatial}
