\name{selectByDistance}

\alias{selectByDistance}

\title{select by distance}

\description{
Select pairs of points from two sets that have the a similar distance to its nearest point in another set.
}

\usage{
selectByDistance(from, to, target, tr=0.1, lonlat=TRUE) 
}
 
\arguments{
  \item{from}{point locations }
  \item{to}{point locations }
  \item{target}{point locations }
  \item{tr}{ arg}
  \item{lonlat}{ Logical}
}

\value{
explain
}

\author{
Robert J. Hijmans
}

\keyword{spatial}
