\name{shp}
\alias{shp}

\title{
read a shapefile
}

\description{
Simple wrapper around readOGR (rgdal package)
}


\usage{
shp(filename)
}
 
\arguments{
  \item{filename}{shapefile name}
}

\value{
Spatial*DataFrame 
}

\author{
Robert J. Hijmans
}

\keyword{spatial}

