\name{dismo-package}
\alias{dismo-package}
\alias{dismo}
\docType{package}
\title{
dismo
}

\description{
more dismo
}
\details{
\tabular{ll}{
Package: \tab dismo\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-12-07\cr
License: \tab GPL 3\cr
LazyLoad: \tab yes\cr
}

~ ~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Robert J. Hijmans
}

\references{

}
\keyword{ package }
\keyword{ spatial }

