\name{dismo-package}

\alias{dismo-package}
\alias{dismo}

\docType{package}

\title{
Species distribution modeling
}

\description{
This package implements a few species distribution models, including an R link to maxent, and native implementations of older models Bioclim and Domain. It also provides a number of funcitons that can assist in using Boosted Regresssion Trees as well as general functions such as sampling background points, k-fold sampling, and evaluation statistics that are useful for these methods other methods available in R such as GLM, GAM, and RandomForest.
}

\details{
\tabular{ll}{
Package: \tab dismo\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 28-Jan-2010\cr
License: \tab GPL 3\cr
LazyLoad: \tab yes\cr
}
}

\author{
Robert J. Hijmans, Steven Phillips, John Leathwick and Jane Elith
}

\keyword{ package }
\keyword{ spatial }

