\name{ecolim}

\alias{ecolim}
\alias{ecolim,matrix,matrix-method}
\alias{EcoLim-class}
\alias{predict,EcoLim-method}


\title{Ecolim model}

\description{
Very simple generic limiting factor based mechanistic model (under development). 
}

\usage{
ecolim(x, y, ...) 
}
 

\arguments{
  \item{x}{Matrix with driver variables}
  \item{y}{Matrix with responses (between 0 and 1)}
  \item{...}{Additinal arguments. None implemented}
}


\author{
Robert J. Hijmans
}

\keyword{spatial}


