\name{evaluate}
\alias{evaluate}

\title{
Model testing
}

\description{
Compute AUC and other statistics.
}

\usage{
evaluate(model, p, a, x, tr)
}
 
\arguments{
  \item{model}{}
  \item{p}{}
  \item{a}{}
  \item{x}{}
  \item{tr}{}
}


\value{
}


\author{
Robert J. Hijmans
}


\keyword{spatial}

