\name{NSW}

\alias{NSWtrain}
\alias{NSWbackground}
\alias{NSWtest}
\alias{NSWeval}

\docType{data}

\title{New South Wales species distribution data}

\description{
NSW (New South Wales, a state in Australia) species distribution data. There are four data sets with training (NSWtrain), background or pseudo-absence (NSWbackground), test (NSWtest), and evaluation (NSWeval) data.
}

\usage{
data(NSWtrain)
data(NSWbackground)
data(NSWtest)
data(NSWeval)
}

\format{

to be done

The format is:
data.frame with xx observations and xx variables:
  \describe{
    \item{Longitude}{numeric. longitude of points}
    \item{Longitude}{numeric. latitude of points}
  }
  
Latitude and latitude are in geographical coordinates using the WGS84 datum.
}

\source{
Data was compiled and provided by Simon Ferrier and colleagues and distributed with his permission. 

These are the NSW data that were used in this model comparison study: 
Elith, J., C.H. Graham, R.P. Anderson, M. Dudik, S. Ferrier, A. Guisan, R.J. Hijmans, F. Huettmann, J. Leathwick, A. Lehmann, J. Li, L.G. Lohmann, B. Loiselle, G. Manion, C. Moritz, M. Nakamura, Y. Nakazawa, J. McC. Overton, A.T. Peterson, S. Phillips, K. Richardson, R. Scachetti-Pereira, R. Schapire, J. Soberon, S. Williams, M. Wisz and N. Zimmerman, 2006. Novel methods improve prediction of species' distributions from occurrence data. Ecography 29: 129-151. \url{http://dx.doi.org/10.1111/j.2006.0906-7590.04596.x}
}

\keyword{datasets}

