\name{plot}

\docType{methods}

\alias{plot,ConfusionMatrix,ANY-method}

\alias{plot,Bioclim,missing-method}

\alias{plot,DistModel,numeric-method}
\alias{points,DistModel-method}


\title{
plot occurence data
}

\description{
.. description ...
}


\section{Methods}{
\describe{
	\code{plot(x, y, ...)}
	\code{points(x, ...)}
}
}


\seealso{
\code{\link[dismo]{density}, \link[dismo]{pairs}}
}

\author{
Robert J. Hijmans
}


\keyword{spatial}

