\name{plot}

\docType{methods}

\alias{plot,Bioclim,missing-method}
\alias{plot,DistModel,numeric-method}
\alias{points,DistModel-method}


\title{
Plot predictor values 
}

\description{
Plot predictor values for occurence (presence and absence) data in a DistModel (or derived) object.
}


\section{Methods}{
usage:  \code{plot(x, y, ...)}
\tabular{rll}{
\tab \code{x}  \tab  Object of class DistModel or from a class that inherits from it \cr
\tab \code{y}  \tab  missing \cr
\tab \code{...}  \tab  Additional arguments that can be passed to \link[graphics]{plot} \cr
}
}


\seealso{
\code{\link[dismo]{density}, \link[dismo]{pairs}, \link[graphics]{plot}}
}

\author{
Robert J. Hijmans
}


\keyword{spatial}

