\name{prepareData}

\docType{methods}

\alias{prepareData}

\title{
Prepare data for model fitting
}

\description{
Simple helper function to prepare data for model fitting
}

\usage{
prepareData(p, b, x, factors) 
}

\arguments{
\item{p}{presence points}
\item{b}{background points}
\item{x}{Raster* object}
\item{factors}{vectors indicating which variables are factors (names or numbers)}
}


\author{
Robert J. Hijmans
}

\keyword{spatial}

