\name{shapefile}
\alias{shapefile}

\title{
read a shapefile
}

\description{
Simple wrapper around readOGR (rgdal package)
}


\usage{
shapefile(filename)
}
 
\arguments{
  \item{filename}{Character. Full filename of a ESRI shapefile}
}

\value{
Spatial*DataFrame 
}

\author{
Robert J. Hijmans
}

\keyword{spatial}

