\name{voronoi}

\Rdversion{1.1}

\alias{voronoi}

\title{
Voronoi polygons
}

\description{
Voronoi polygons
}

\usage{
voronoi(xy)
}

\arguments{
 \item{xy}{SpatialPoints* or two column matrix}
}

\value{
SpatialPolygonsDataFrame
}

\author{
Adapted from code by Carson Farmer: \url{http://www.carsonfarmer.com/?p=455}
}

\keyword{ spatial  }
